Pre-filter the text lines for 'Camera Temperature'
This is a performance improvement patch.
- By pre-filtering the lines that are checked by the regex we reduce the
scaling impact of any changes to the regex to almost zero (like checking
for temperatures <10 or >99 or using a quantifier).
- Changing to readlines() instead of bare read() lets us make this
change. Notably, this improves performance in the case of processing
logs that can be opened as native text files (.DM.log files) by a factor
of 3-4x. Performance for zip files is unchanged, because the performance
impact of the regex is replaced by the overhead from the zipfile
.readline method. There may be additional optimization, but that remains
to be sorted out in a future patch.